+++ src/renderer/thumbnail/composables/useWorkflowThumbnail.ts
